[#3636] <OrganizationSelect>; /partners refactor
