+++ akvo/rsr/spa/app/utils/hooks.js
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react'
-import api from '../../utils/api'
+import api from './api'
 
 function useFetch(url) {
   const [data, setData] = useState([])
